requirements.txt
python-docx
spacy==2.3.5
PyMuPDF==1.16.14
jsonlines
titlecase
